+++ internal/pproflogging/pproflogging_test.go
@@ -1,4 +1,4 @@
-package debug
+package pproflogging
 
 import (
 	"fmt"
